group0_state_machine: flush system.topology when updating features columns
The `supported_features` and `enabled_features` columns from
`system.topology` are read during the feature check that happens early
on boot. The check enforces two properties:

- A node is not allowed to revoke support for a feature after it notices
  in its local topology state that the feature is supported by all
  nodes.
- Similarly, a node is not allowed to revoke support for a feature after
  seeing that it was put to the `enabled_features` column by the
  topology coordinator.

However, due to the fact that the check has to happen before (schema)
commitlog replay and the table is not explicitly flushed when
`supported_features` or `enabled_features` columns are modified, the
feature check on boot might operate on old data and not do its job
properly.

In order to fix this, this commit modifies the `group0_state_machine` so
that is flushes the `system.topology` table every time the
`supported_features` or `enabled_features` column is modified, and after
every snapshot transfer.