topology_coordinator: remove barrier_after_feature_update
The `barrier_after_feature_update` was introduced as a variant of the
`barrier` command, meant to be used by the topology coordinator when
enabling a feature. It was meant to give more guarantees to the topology
coordinator than the regular barrier, but the regular barrier has been
adjusted in the previous commits so that it can be used instead of the
special barrier.

This commit gets rid of `barrier_after_feature_update` and replaces its
uses with `barrier`.